requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.1
aiofiles==24.2.0
asyncpg==0.29.0
asyncio-mqtt==0.16.2
Pillow==10.4.0
